% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{create_raster_inputs}
\alias{create_raster_inputs}
\title{Creates raster stack of coastline, overwater distance, and bathymetry}
\usage{
create_raster_inputs(
  obs_center,
  max_range,
  bathymetry,
  analysis_crs,
  raster_res = 500
)
}
\arguments{
\item{obs_center}{Coordinates (decimal degrees) of central place, from which
over water distances will be calculated}

\item{max_range}{Maximum distance (m) from the central place the species is
expected to travel.}

\item{bathymetry}{raster where depth values should be meters below 0,
used to define coastline}

\item{analysis_crs}{a proj4string for the projection of raster outputs, should have units in m}
}
\description{
Creates raster stack of coastline, overwater distance, and bathymetry
}
