% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_trips}
\alias{sim_trips}
\title{Uses a particle filter to simulate trips based on observed movement rates,
conditioned on observed PDF of depth and distance from colony}
\usage{
sim_trips(
  data,
  origin_col,
  sim_col,
  dist_origin,
  dist_sim,
  water_origin,
  water_sim,
  bath_origin,
  bath_sim,
  n_trips = 50,
  samps = 1000
)
}
\description{
Uses a particle filter to simulate trips based on observed movement rates,
conditioned on observed PDF of depth and distance from colony
}
